requirements.txt
# requirements.txt

Flask
colormap
easydev
qrcode
Pillow
CustomTkinter

